fixed GetTime

+++ Tabby/src/Tabby/Utils/PlatformUtils.h
@@ -4,18 +4,16 @@
 
 namespace Tabby {
 
-	class FileDialogs
-	{
-	public:
-		// These return empty strings if cancelled
-		static std::string OpenFile(const char* filter);
-		static std::string SaveFile(const char* filter);
-	};
+class FileDialogs {
+public:
+    // These return empty strings if cancelled
+    static std::string OpenFile(const char* filter);
+    static std::string SaveFile(const char* filter);
+};
 
-	class Time
-	{
-	public:
-		static float GetTime();
-	};
+class Time {
+public:
+    static double GetTime();
+};
 
 }
